refactor(tests): consolidate Installer Module — 11 files → 2
Closes #3758
